requirements.txt
dotenv
streamlit
langchain_google_genai
langchain
langchain_community
google.generativeai
pathlib
